\name{make}
\alias{make}
\title{
Small helper function.
}
\description{
Helps sq.pe.
}
\usage{
make(r1, r2, k, v)
}
\arguments{
  \item{r1}{
lower bound}
  \item{r2}{
upper bound
}
  \item{k}{
internal counter
}
  \item{v}{
number of columns}
}
\details{
helper function for sq.pe()
}
\value{
a vector}
\references{
none.
}
\author{
Eric Golinko}
\note{
thank you.
}
\examples{
#sq.pe(r1,r2,v)}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manpi }

