\name{seqPERM-package}
\alias{seqPERM-package}
\alias{seqPERM}
\docType{package}
\title{
Generates a permutation matrix based upon a sequence.
~~ seqPERM ~~
}
\description{
User will enter a lower and upper bound and how many columns are desired.

}
\details{
\tabular{ll}{
Package: \tab seqPERM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-28\cr
License: \tab GPL-3\cr
}
sq.pe(r1,r2,v)
}
\author{
Eric Golinko

Maintainer: Eric Golinko <egolinko@gmail.com>

}
\references{
none.
}

\keyword{ package }
\seealso{
none.}
\examples{
#sq.pe(0,1,3)
}
