\name{sq.pe}
\alias{sq.pe}
\title{
Function that carries out generating the permutation matrix.}
\description{
Enter lower and upper bound of a sequence to generate a permuation matrix.
}
\usage{
sq.pe(r1, r2, v)
}

\arguments{
  \item{r1}{
  	     lower bound of sequence 
}
  \item{r2}{
     upper bound of sequence
}
  \item{v}{
     number of columns to generate }
}
\details{
Use function to generate matrix.
}
\value{
A matrix.
}
\references{
none.
}
\author{
Eric Golinko}
\note{
thank you.
}

\examples{
#sq.pe(0,1,3)
}    

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manpi }

